Validate disease terms and don't drop old terms before validation ends (#4114)
* Turn disease into a pydantic class

* Isorted file

* Work in progress

* Rename function in other file

* Some fixes

* Some more fixes

* Updated changelog

* Do no drop old diseases without checking that the new ones are parsed correctly

* Thanks sonarcloud

* Drop old collection later

* Validate the disease_nr before

* Reintroduce files check

* Moved load_hpo to a separete module

* Work in preogress

* Small fixes

* Small but important fixes

* Remove repeated lines

* Remove old DiiseaseTerm

* Fix model and validation type of error in tests

* Include a commented line explaining that the database object has an additional _id field

* Fix comment

* Remove load phenotype

---------
