+++ scout/load/__init__.py
@@ -3,6 +3,5 @@ from .case import load_case
 from .cytoband import load_cytobands
 from .exon import load_exons
 from .hgnc_gene import load_hgnc_genes
-from .hpo import load_hpo
 from .institute import load_institute
 from .transcript import load_transcripts
