+++ scout/models/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from .institute import Institute
-from .phenotype_term import DiseaseTerm, HpoTerm
+from .phenotype_term import HpoTerm
 from .user import User
